* 3547_cleanup: (45 commits)
  Update po/*.po.
  Update copyright years.
  Robust sizeof() usage at function parameter and use memcpy destination sizeof...
  Remove unused function exec_shell().
  viewer: remove unused const value.
  diffviewer: use gboolean at WDiff struct.
  Clarify function declare at USE_DIFF_VIEW.
  (tree_store_load_from): optimise and fix return value.
  (tree_store_save): optimisation.
  cleanup gcc link time optimization warnings
  (tree_entry, TreeStore): use gboolean instead of bit fields.
  (edit_sort_cmd): fix memory leak.
  Ticket #3572: (edit_set_spell_lang): fix memory leak.
  (configure_panel_listing): fix memory leak.
  Ticket #3567: fix heap-use-after-free bug when accessing already freed widget object
  Use STR_E_NOTFOUND message where required.
  (listbox_key): remove redundant code.
  (extfs_find_entry_int): fix comparisons.
  (vfs_s_print_stats): avoid macros in i18n resources.
  Use signed long instead of unsigned long for binded actions.
  ...

+++ lib/widget/menu.h
@@ -34,7 +34,7 @@ typedef struct WMenuBar
     gboolean is_active;         /* If the menubar is in use */
     gboolean is_dropped;        /* If the menubar has dropped */
     GList *menu;                /* The actual menus */
-    size_t selected;            /* Selected menu on the top bar */
+    guint selected;             /* Selected menu on the top bar */
     unsigned long previous_widget;      /* Selected widget ID before activating menu */
 } WMenuBar;
 
@@ -42,7 +42,7 @@ typedef struct WMenuBar
 
 /*** declarations of public functions ************************************************************/
 
-menu_entry_t *menu_entry_create (const char *name, unsigned long command);
+menu_entry_t *menu_entry_create (const char *name, long command);
 void menu_entry_free (menu_entry_t * me);
 
 menu_t *create_menu (const char *name, GList * entries, const char *help_node);
